+++ patzilla-ui/navigator/app/document/index.js
@@ -1,0 +1,3 @@
+require('./document-controller.js');
+require('./document-view.js');
+require('./family.js');