+++ patzilla-ui/navigator/app/results/index.js
@@ -1,0 +1,2 @@
+require('./results-header.js');
+require('./results-list.js');