+++ CHANGES.rst
@@ -14,6 +14,7 @@ development
 
     - Data source adapters
     - Application layout- and error templates
+    - Result list and document details
 
 2017-10-06 0.137.0
 ==================